+++ packages/rath-client/src/store/commonStore.ts
@@ -2,7 +2,7 @@ import { makeAutoObservable, observable, runInAction } from 'mobx';
 import { Specification } from 'visual-insights';
 import { COMPUTATION_ENGINE, EXPLORE_MODE, PIVOT_KEYS } from '../constants';
 import { IAccessPageKeys, ITaskTestMode, IVegaSubset } from '../interfaces';
-import { getAvatarURL, getServerUrl, AVATAR_IMG_LIST, IAVATAR_TYPES } from '../utils/user';
+import { getAvatarURL, getMainServiceAddress, AVATAR_IMG_LIST, IAVATAR_TYPES } from '../utils/user';
 import { destroyRathWorker, initRathWorker, rathEngineService } from '../services/index';
 import { transVegaSubset2Schema } from '../utils/transform';
 import { notify } from '../components/error';
@@ -162,7 +162,7 @@ export class CommonStore {
     }
 
     public async liteAuth(certMethod: 'email' | 'phone') {
-        const url = getServerUrl('/api/liteAuth');
+        const url = getMainServiceAddress('/api/liteAuth');
         const { certCode, phone, email } = this.signup;
         const res = await fetch(url, {
             method: 'POST',
@@ -215,7 +215,7 @@ export class CommonStore {
 
     public async commitLogout() {
         try {
-            const url = getServerUrl('/api/logout');
+            const url = getMainServiceAddress('/api/logout');
             const res = await fetch(url, {
                 method: 'GET',
             });
@@ -240,7 +240,7 @@ export class CommonStore {
 
     public async updateAuthStatus() {
         try {
-            const url = getServerUrl('/api/loginStatus');
+            const url = getMainServiceAddress('/api/loginStatus');
             const res = await request.get<{}, { loginStatus: boolean; userName: string }>(url);
             if (res.loginStatus && res.userName !== null) {
                 runInAction(() => {
@@ -257,7 +257,7 @@ export class CommonStore {
         }
     }
     public async getPersonalInfo() {
-        const url = getServerUrl('/api/ce/personal');
+        const url = getMainServiceAddress('/api/ce/personal');
         try {
             const result = await request.get<{}, IUserInfo>(url);
             if (result !== null) {
@@ -282,7 +282,7 @@ export class CommonStore {
             const { file } = value;
             const data = new FormData();
             file && data.append('file', file);
-            const url = getServerUrl('/api/ce/avatar');
+            const url = getMainServiceAddress('/api/ce/avatar');
             const res = await fetch(url, {
                 method: 'POST',
                 credentials: 'include',
@@ -303,7 +303,7 @@ export class CommonStore {
     }
 
     public async getAvatarImgUrl() {
-        const url = getServerUrl('/api/ce/avatar');
+        const url = getMainServiceAddress('/api/ce/avatar');
         try {
             const result = await request.get<{}, { avatarURL: string }>(url);
             if (result !== null) {