+++ packages/rath-client/src/utils/user.ts
@@ -1,3 +1,5 @@
+import { RathEnv, RATH_ENV } from "../constants";
+
 export enum IAVATAR_TYPES {
     gravatar = 'gravatar',
     default = 'default'
@@ -16,12 +18,16 @@ export const AVATAR_IMG_LIST: string[] = new Array(18)
 
 export const DEFAULT_AVATAR_URL_PREFIX = 'https://foghorn-assets.s3.ap-northeast-1.amazonaws.com/avatar/';
 
-export function getServerUrl(path: string) {
+const DEFAULT_MAIN_SERVER_HOST = `${
+    window.location.host.match(/kanaries\.[a-z]+$/i)?.[0] ?? 'kanaries.net'
+}`;
+
+export function getMainServiceAddress(path: string) {
     const baseURL = new URL(window.location.href);
-    const DATA_SERVER_URL =
-        baseURL.searchParams.get('main_service') || localStorage.getItem('main_service') || window.location.href;
-    // const devSpecURL = new URL(w|| window.location.href)
-    const url = new URL(DATA_SERVER_URL);
+    const CONFIGURABLE_MAIN_SERVER_URL = RathEnv === RATH_ENV.ONLINE ? null
+        : baseURL.searchParams.get('main_service') || localStorage.getItem('main_service');
+    const serverUrl = CONFIGURABLE_MAIN_SERVER_URL ?? `${baseURL.protocol}//${DEFAULT_MAIN_SERVER_HOST}`;
+    const url = new URL(serverUrl);
     url.pathname = path;
     return url.toString();
 }