+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 > Released on Nov.23.2022
 The initial version of RATH changelog is released. Incoming updates will be grouped in the following categories:
 
-### Feature Requests
+#### Feature Requests
 - Support more types of data file. [#150](https://github.com/Kanaries/Rath/issues/150)
   - base: txt with customized splitter, xlsx file
   - cool : multiple-layer JSON data with LaTiao transformation.
@@ -20,7 +20,21 @@ The initial version of RATH changelog is released. Incoming updates will be grou
   - search pill
 - support cases when datasets contains no measures [#95](https://github.com/Kanaries/Rath/issues/95)
 
-### Bug fixes
+#### Bug fixes
 - [Vega-Lite] This file will be used when vega-lite facets bug is fixed. [#70](https://github.com/Kanaries/Rath/issues/70)
 
 [Back to TOC](##table-of-contents)
+
+
+### 1.1.0
+
+#### Features
++ Causal analysis
+  + causal discovery
+  + ML modeling
+  + causal explain
++ Text pattern discovery: User can specify a selection in the data table, RATH can figure out the selection purpose and genelize a pattern to match all same patterns and extract to new variable.
++ Data Table do not display axis' labels and use tooltip instead.
++ image download button beside visualization.
++ better file upload UI
++ support Excel File as DataSource.